==> updated-notes/3-coerce-dupes-and-dates.ipynb <==
# Identify DataTypes

### Introduction

### Sizing it Up

> You can find this dataset [here](https://github.com/jigsawlabs-student/engineering-large-datasets/blob/master/X_train_listings.feather).  You need to download by clicking *view raw*, and then upload to your colab notebook.

import pandas as pd 

X_train_df = pd.read_csv('./price_listings_ten_k.csv', index_col = 0)

def find_all_same(df):
    return [col for col in df.columns if len(df[col].unique()) == 1]

same_cols = find_all_same(X_train_df)

same_cols

X_train_df[same_cols][:2]

X_train_df_var = X_train_df.drop(columns = same_cols)

X_train_df_var.shape

So right off the bat, we got rid of some columns.  Now let's further reduce our problem by finding those columns that are of type object.  These will be the ones that we'll need to work on coercing.

X_train_object_df = X_train_df_var.select_dtypes(include = 'object')

X_train_object_df.shape

So we see that there are 29 columns that we can effectively use and 56 that we cannot use.

### Coerce To DateTimes

Now let's identify and coerce our date columns.  We first use our method to identify the date columns.

def contains_date(column):
#     remove nas first, potentially use all
    regex_string = (r'^\d{1,2}-\d{1,2}-\d{4}$|^\d{4}-\d{1,2}-\d{1,2}$' + 
'|^\d{1,2}\/\d{1,2}\/\d{4}$|^\d{4}\/\d{1,2}\/\d{1,2}$')
    return column.str.contains(regex_string).any()

potential_date_cols = [col for col in X_train_object_df.columns if contains_date(X_train_object_df[col])]

potential_date_cols

Then we take a look.

almost_date_df = X_train_object_df[potential_date_cols]
almost_date_df[:2]

We have identified five columns that we can set as datetimes.  Let's do so.


date_df = almost_date_df.apply(lambda col: pd.to_datetime(col, errors = 'coerce'))
date_df[:3]

Now we can replace our original dataframe with these columns.

X_train_df_var.loc[:, date_df.columns] = date_df

So we can see that we just coerced five more columns.

X_train_df_var.select_dtypes(include = 'object').shape

So we just finished with coercing to dates.  It wasn't so bad.  We just needed to identify the datecolumns, and pass a datatype dictionary to the `astype` method.

### Coercing Numbers

Let's move onto seeing if there are any columns that we can convert to numbers.  We start off in a similar way as before.  We again select the columns that are of type object, and then see which of those appear to contain numbers.

X_train_object_df = X_train_df_var.select_dtypes(include = 'object')

> Now that we narrowed down to object columns, we can see which ones contain numbers.

To do so, we have a function that looks through a column and determines if there are any numbers.

def contains_numbers(column):
    # matches price or percentage     
    regex_string = (r'^(?!.*www|.*-|.*\/|.*[A-Za-z]|.* ).*\d.*')
#     regex_string = (r'\$\d+.*|\d+.*\%$|^\d+.*$')
    return column.str.contains(regex_string).all()

> We can change the function to end with `any` instead of `all` to make it less restrictive.

Then, we loop through our columns to see which contain numbers.

potential_num_cols = [col for col in X_train_object_df.columns if contains_numbers(X_train_object_df[col])]

potential_num_cols

Let's take a look.

almost_num_df = X_train_object_df[potential_num_cols]
almost_num_df[:6]

Ok, so our function did a pretty good job.

Ok now that we've identified our almost number columns, let's work on coercing them.  

We already have `remove_price` method, which seems like it will work on our `security_deposit`, `cleaning_fee` and `extra_people` columns.

> One thing to consider is whether these last three columns really belong in our features.  The columns  are related to how much our host is charging for the apartment, which is what we are trying to predict.  Let's include it for now, but this may be something to remove later on.

Ok, we can start by selecting a single column, and then working on converting it.  We'll start with the `host_response_rate`.

host_resp_rate = almost_num_df['host_response_rate']

host_resp_rate.loc[5]

Then we can write a function that removes a percentage where one exists, simply converts null values to be numeric.

import pandas as pd
def convert_percent(val):
    if pd.isna(val):
        return pd.to_numeric(val, errors='coerce', downcast='float')
    else:
        without_dollar = val[:-1]
        return float(without_dollar)

We can test that this works properly with the map function.

host_resp_rate.map(convert_percent)[:3]

And then we can use the function with a FunctionTransformer.

from sklearn_pandas import FunctionTransformer, DataFrameMapper

from sklearn.impute import SimpleImputer
mapper = DataFrameMapper([
    (['host_response_rate'], FunctionTransformer(convert_percent))
], df_out = True)

transformed_host_rate = mapper.fit_transform(almost_num_df[['host_response_rate']])

transformed_host_rate[:3]

Great, so now we wrote a DataFrameMapper to convert this into a numeric column.

transformed_host_rate.dtypes

Let's move onto the rest of the columns.

almost_num_df[:6]

We'll convert the price columns in the same way. 

def convert_price(val):
    if pd.isnull(val):
        return pd.to_numeric(val) 
    else:
        without_commas = val.replace(',','')
        without_dollar = without_commas[1:]
        return pd.to_numeric(without_dollar)

> Then we can place this integrate this function into a step with a FunctionTransformer.

(['security_deposit'], [FunctionTransformer(convert_price)])

But remember, we have three columns that we wish to convert.

potential_num_cols[1:]

So we can loop through and create a list of the steps with code.

convert_dolls_to_nums = [([col], [FunctionTransformer(convert_price)]) 
                         for col in potential_num_cols[1:]]

convert_dolls_to_nums

Take a look at the code above.  We'll be doing this a lot.  All we're doing is looping through to create multiple steps with list iteration.  So we specify the each step that we want to return as:
```python
([col], [FunctionTransformer(convert_price)])
```

Once our steps are defined, we integrate them in our DataFrameMapper.

from sklearn.impute import SimpleImputer
convert_percent_to_nums = [ (['host_response_rate'], [FunctionTransformer(convert_percent)]) ]
convert_dolls_to_nums
convert_to_nums_steps = convert_percent_to_nums + convert_dolls_to_nums
to_number_mapper = DataFrameMapper(convert_to_nums_steps, df_out = True, default = None)

converted_nums_df = to_number_mapper.fit_transform(X_train_df_var)

converted_nums_df[:1]

So things look pretty good.  We added default = None, so that we return the entire dataframe.

Unfortunately, the `default = None`, changed our unspecified columns to be type Object.  No worry, we change them back.  We do so by collecting the dtypes from our dataframe above.  And then only selecting those datatypes that were of type object.

dtypes = dict(X_train_df_var.dtypes)

dtypes = {k:v for k, v in dtypes.items() if v.kind is not 'O'}

list(dtypes.items())[:5]

updated_df['last_scraped'].dtype

Because these are the dtypes that we want to keep, we can then update our dataframe with them.

updated_df = converted_nums_df.astype(dtypes)

Now we can see that we have 38 columns in the correct format.  Not to bad.

updated_df.select_dtypes(exclude = 'object').shape

updated_df.to_csv('./nums_and_dates_ten_k.csv')

### Summary

In this lesson, we worked through our dataset to convert columns to be of type datetime and numeric columns.  With datetimes, we used apply to loop through our almost date columns, and convert each to a datetime, and then updated our dataframe.  

With the numeric columns we again identified the almost numeric columns, then we wrote functions to coerce them into numbers, and integrated the functions into our FunctionTransformer to use with our DataFrameMapper.
