==> updated-notes/6-coerce-numbers-lab.ipynb <==
# Numerical Data Lab

## Introduction
In this lab we will use the feature lib to transform the numerical data in the `car_data.csv` dataset. Time to start.

import pandas as pd
url = "https://raw.githubusercontent.com/jigsawlabs-student/engineering-large-datasets/master/car_data.csv"
df = pd.read_csv(url)

df[:2]

We can see that two of the columns have numeric data in them, `Selling_Price` and `Present_Price`.  Let's just select them, as we have a small enough dataset to spot them.

almost_nums_df = df[['Selling_Price', 'Present_Price']]

Ok, now let's use a list comprehension to create our steps, and then coerce the data with a DataFrameMapper. 

> Try not to reference the previous reading at first.  Only look to it if you get stuck.

import pandas as pd
def price_to_num(val):
    pass

steps = None

mapper = None

transformed_cols = mapper.fit_transform(df)
transformed_cols[:5]

# 	Selling_Price	Present_Price
# 0	3.35	5.59
# 1	4.75	9.54
# 2	7.25	9.85
# 3	2.85	4.15
# 4	4.60	6.87

### Working with Year

Ok, now let's add year in there.  We can write a method that coerces the year, and use it with a transformer to add to our mapper.  Let's get going.

def coerce_to_year(val):
    pass

Store the step in `coerce_step`.

coerce_step = None

Then create a list of steps that convert both the prices and the year.

comb_steps = None

And add the list of steps to the mapper.

mapper_with_num_converter = None

price_year_df = mapper_with_num_converter.fit_transform(df)

price_year_df[:3]

# 	Selling_Price	Present_Price	Year
# 0	3.35	5.59	6
# 1	4.75	9.54	7
# 2	7.25	9.85	3

### Keeping the rest

coerced_df.dtypes

We can see that we lost our original int datatypes from our starting dataframe. 

df.dtypes

So below, we'll select the datatypes from `df`.  

df_dtypes = df.dtypes.to_dict()
# df_dtypes = {'Car_Name': dtype('O'),
#  'Year': dtype('O'),
#  'Selling_Price': dtype('O'),
#  'Present_Price': dtype('O'),
#  'Kms_Driven': dtype('int64'),
#  'Fuel_Type': dtype('O'),
#  'Seller_Type': dtype('O'),
#  'Transmission': dtype('O'),
#  'Owner': dtype('int64')}

Then use dictionary comprehension to select those that are not of type object.

import numpy as np
non_obj_dtypes = None

non_obj_dtypes
# {'Kms_Driven': dtype('int64'), 'Owner': dtype('int64')}

Then use this dictionary to update our datatypes of the `coerced_df`.

updated_df = None

updated_df.dtypes

# Selling_Price    float64
# Present_Price    float64
# Car_Name          object
# Year              object
# Kms_Driven         int64
# Fuel_Type         object
# Seller_Type       object
# Transmission      object
# Owner              int64
# dtype: object

### Summary

In this lesson, we worked with coercing our numeric data.  We practiced using list iteration to create multiple steps simultaneously, and also worked with coercing our datatypes using a dtypes dictionary.
