==> updated-notes/7-coerce-to-boolean.ipynb <==
# Coercing to Booleans

1. GEt the object df

# df

object_df = df.select_dtypes('object')

# [col for col in object_df.columns if is_date_time(object_df[col])]

2. Identify columns to change similarly

# almost_datetime_cols = object_df.apply(lambda col: is_date_time(col))

3. Loop through and make the change

# steps = [ ( [almost_dt_col], FunctionTransformer(change_it_to_date)) for almost_dt_col in almost_datetime_cols]

from sklearn_pandas import DataFrameMapper

mapper = DataFrameMapper(steps, df_out = True)
mapper.fit_transform()

# columns that we coerced

### Loading our AirBnb Data

For this lesson, we'll work with [AirBnb listings in Berlin](https://www.kaggle.com/brittabettendorf/berlin-airbnb-data).  Let's load our data.

import pandas as pd
df = pd.read_csv('./nums_and_dates_ten_k.csv', index_col = 0)

potential_date_cols = ['last_scraped',
 'host_since',
 'calendar_last_scraped',
 'first_review',
 'last_review']
df[potential_date_cols] = df[potential_date_cols].astype('datetime64')

def contains_date(column):
#     remove nas first, potentially use all
    regex_string = (r'^\d{1,2}-\d{1,2}-\d{4}$|^\d{4}-\d{1,2}-\d{1,2}$' + 
'|^\d{1,2}\/\d{1,2}\/\d{4}$|^\d{4}\/\d{1,2}\/\d{1,2}$')
    return column.str.contains(regex_string).any()

### Feature engineering

.value_counts()

df['last_scraped'].unique()

def percent_different(df_series):
    series_filled = df_series.dropna()
    return len(series_filled.unique())/len(series_filled)

# percent_different(df['last_scraped'])

def find_categorical(df, threshold = .5):    
    categorical_df = pd.DataFrame({})
    for column in df.columns:
        if percent_different(df[column]) < threshold:
            categorical_df[column] = df[column]
    return categorical_df 

cat_df = find_categorical(df.select_dtypes('object'))

# cat_df

### Combine with Selecting Categorical Columns

def get_multiple_val_counts(df, num_vals = 1):
    return [df[column].value_counts(normalize=True).iloc[:num_vals] for column in df.columns]

# get_multiple_val_counts(cat_df, 2)

import numpy as np
def summarize_cats(df):
    multiple_val_counts = get_multiple_val_counts(df)
    stacked_counts = np.vstack([np.array([val_count.name, val_count.index[0], float(val_count.values[0])]) for val_count in multiple_val_counts])
    sorted_cols = np.argsort(stacked_counts.reshape(-1, 3)[:, 2].astype('float'))
    return stacked_counts[sorted_cols[::-1]]

summary = summarize_cats(cat_df)

summary

import numpy as np
bool_cols = summary[np.isin(summary[:, 1], ['t', 'f'])]
bool_cols[:, 0]

bool_df = cat_df[bool_cols[:, 0]]

# steps = [([col], [SimpleImputer(strategy='constant', fill_value='f'),MissingIndicator(missing_values='t')) for col in bool_cols[:, 0]]

# get_multiple_val_counts(bool_df, 2)

# bool_df

from sklearn.impute import MissingIndicator, SimpleImputer
missing_indicator_steps = [([col], [SimpleImputer(strategy='constant', fill_value = 'f'), MissingIndicator(missing_values='t')])for col in bool_cols[:, 0]]

from sklearn_pandas import DataFrameMapper


mapper = DataFrameMapper(missing_indicator_steps, df_out = True)

bool_df = mapper.fit_transform(bool_df)

# bool_df

df_updated_bool = df.drop(columns = bool_df.columns)

len(bool_df.columns)

df_updated_bool[bool_df.columns] = bool_df

df_updated_bool.select_dtypes('object').shape

df.select_dtypes('object').shape

### Identifying Boolean Values
