==> updated-notes/9-coerce-to-categorical.ipynb <==
# Coercing to Categorical

### Introduction

### Loading our AirBnb Data

1. Go after the low hanging fruit 
1. Numeric 
        * price and percent
2. Datetimes
3. Booleans
4. Categorical 
5. Datetimes-> add_datepart (weekday)
6. Get rid of our nas -> impute, is_na column

We'll start by loading the data where we last left off.

import pandas as pd
df = pd.read_csv('./coerced_bools_complete.csv')

And let's load our datatypes and set our them in our dataframe.

import requests 
url = "https://raw.githubusercontent.com/jigsawlabs-student/engineering-large-datasets/master/coerced_bools_complete_dtypes.json"
response = requests.get(url)
dtypes = response.json()

# df.to_feather('./whatever.feather')

df = df.astype(dtypes)

object_df = df.select_dtypes('object')

# object_df

### Loading our Library

def get_multiple_val_counts(df, num_vals = 1):
    return [df[column].value_counts(normalize=True).iloc[:num_vals] for column in df.columns]

import numpy as np
def summarize_cats(df):
    multiple_val_counts = get_multiple_val_counts(df)
    stacked_counts = np.vstack([np.array([val_count.name, val_count.index[0], float(val_count.values[0])]) for val_count in multiple_val_counts])
    sorted_cols = np.argsort(stacked_counts.reshape(-1, 3)[:, 2].astype('float'))
    return stacked_counts[sorted_cols[::-1]]

summarize_cats(object_df)[:10]

### Selecting Columns

Then we can use the `selected_cat_values` method to take a deeper look at the values in each of the columns.

def selected_cat_values(column, threshold = .02):
    values_counted = column.value_counts(normalize=True)
    return values_counted[values_counted > threshold]

# [selected_cat_values(object_df[col]) for col in object_df.columns][:20]

### Coercing our Categorical Columns

def reduce_cat_values(column, threshold = .02):
    column = column.copy()
    selected_values = selected_cat_values(column, threshold).index
    column[~column.isin(selected_values)] = 'other'
    column.astype('category')
    return column

# FunctionTransformer()

cat_cols = summarize_cats(object_df)[:16, 0]
cat_cols

cat_df = object_df[cat_cols]
cat_df.shape

reduced_cat_df = cat_df.apply(lambda col: reduce_cat_values(col))

# [for col in reduced_cat_df.columns]

# get_multiple_val_counts(reduced_cat_df, num_vals = 3)

### Integrating our Mapper

reduced_cat_df.columns

from sklearn.preprocessing import OneHotEncoder
steps = [([col], OneHotEncoder() )for col in reduced_cat_df.columns]

from sklearn_pandas import DataFrameMapper

mapper = DataFrameMapper(steps, df_out = True)

transformed_cat = mapper.fit_transform(reduced_cat_df)

transformed_cat.shape

coerced_cat_df = mapper.fit_transform(reduced_df)

cat_cols = reduced_df.columns

### Aggregating our Data

At this point, we can take our original dataframe.

import pandas as pd
df = pd.read_csv('./coerced_bools_complete.csv', index_col = 0)

import json

file = './coerced_bools_complete_dtypes.json'
with open(file, 'r') as f:
    dtypes = json.load(f)

df = df.astype(dtypes)

df.shape

And drop the categorical columns, and add in our `coerced_cat_df`.

df_dropped_cats = df.drop(columns = cat_cols)

df_dropped_cats[coerced_cat_df.columns] = coerced_cat_df

Let's take a look at what object columns we have left.

remaining_object_df = df_dropped_cats.select_dtypes('object')

remaining_object_df.shape

remaining_object_df[:2].T

It looks  like these are not categorical (with the exception perhaps of amenities and licence), so let's export our data.

df_dropped_cats.to_csv('./coerced_cats.csv')

dtypes = df_dropped_cats.dtypes.astype(str).to_dict()

import json
with open('./dtypes_coerced_cats.json', 'w') as f:
    json.dump(dtypes, f)

### Summary 

In this lesson we coercing our categorical data by first identifying our categorical features with the `summarize_cats` method.  We then used `selected_cat_values` to take a look at the common values in each of these columns. 

Finally, we moved onto coercing our categorical columns.  We made three coercions in all.  First, we replaced sparse values with `other`.  Then, we replaced na values and applied one hot encoding with a DataFrameMapper.
